pyproject.toml
[project]
name = "auth-project"
version = "0.1.0"
description = "Authentication with social accounts"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework",
    "django-allauth[socialaccount]",
    "dj-rest-auth",
    "django-jazzmin",
    "python-dotenv>=1.2.2,<2.0.0",
    "celery>=5.6.2,<6.0.0",
    "djangorestframework-simplejwt>=5.5.1,<6.0.0",
    "dj-database-url>=2.1.0,<3.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
